The beginning of a series of updates to get the FOGI analysis working with many-qubit models.
This commit adds implementation surrounding a (new) ImplicitOpModel.setup_fogi method, which
so far is specific to local noise models and will probably be moved there once we start
working with the cloud noise models.  Additional supporting functionality is also added:
- adds ok_if_missing argument to ElementaryErrorgenBasis.label_indices & label_index methods.
- adds the (existing) _NoErrorGeneratorInterface class as a base class of static SPAM operators.
- moves a bunch of methods that should work for both implicit and explicit models from ExplicitOpModel
  to OpModel.

We've begun testing this using LocalNoiseModel objects, and know that the FOGI analysis
isn't working correctly quite yet, even though it is runnable -- a WIP!
