+++ pygsti/models/model.py
@@ -30,6 +30,7 @@ from pygsti.baseobjs.basis import Basis as _Basis
 from pygsti.baseobjs.label import Label as _Label
 from pygsti.baseobjs.resourceallocation import ResourceAllocation as _ResourceAllocation
 from pygsti.tools import slicetools as _slct
+from pygsti.tools import matrixtools as _mt
 
 MEMLIMIT_FOR_NONGAUGE_PARAMS = None
 
@@ -1680,6 +1681,334 @@ class OpModel(Model):
         mmg = self.create_modelmember_graph()
         other_mmg = other_model.create_modelmember_graph()
         return mmg.is_equivalent(other_mmg, rtol, atol)
+
+    def _format_gauge_action_matrix(self, mx, op, reduce_to_model_space, row_basis, op_gauge_basis,
+                                    create_complete_basis_fn):
+
+        from pygsti.baseobjs.errorgenbasis import CompleteElementaryErrorgenBasis as _CompleteElementaryErrorgenBasis
+        from pygsti.baseobjs.errorgenbasis import ExplicitElementaryErrorgenBasis as _ExplicitElementaryErrorgenBasis
+        from pygsti.baseobjs.errorgenspace import ErrorgenSpace as _ErrorgenSpace
+        import scipy.sparse as _sps
+
+        #Next:
+        # - make linear combos of basis els so that all (nonzero) disallowed rows become zero, i.e.,
+        #    find nullspace of the matrix formed from the (nonzero) disallowed rows of mx.
+        # - promote op_gauge_basis => op_gauge_space (linear combos of the same elementary basis - op_gauge_basis)
+        all_sslbls = self.state_space.tensor_product_block_labels(0)
+
+        if reduce_to_model_space:
+            allowed_lbls = op.errorgen_coefficient_labels()
+            allowed_lbls_set = set(allowed_lbls)
+            allowed_row_basis = _ExplicitElementaryErrorgenBasis(self.state_space, allowed_lbls, basis1q=None)
+            disallowed_indices = [i for i, lbl in enumerate(row_basis.labels) if lbl not in allowed_lbls_set]
+
+            if len(disallowed_indices) > 0:
+                disallowed_rows = mx[disallowed_indices, :]  # a sparse (lil) matrix
+                allowed_gauge_linear_combos = _mt.nice_nullspace(disallowed_rows.toarray(), tol=1e-4)  # DENSE for now
+                mx = _sps.csr_matrix(mx.dot(allowed_gauge_linear_combos))  # dot sometimes/always returns dense array
+                op_gauge_space = _ErrorgenSpace(allowed_gauge_linear_combos, op_gauge_basis)  # DENSE mxs in eg-spaces
+                #FOGI DEBUG: print("DEBUG => mx reduced to ", mx.shape)
+            else:
+                op_gauge_space = _ErrorgenSpace(_np.identity(len(op_gauge_basis), 'd'), op_gauge_basis)
+        else:
+            allowed_row_basis = create_complete_basis_fn(all_sslbls)
+            op_gauge_space = _ErrorgenSpace(_np.identity(len(op_gauge_basis), 'd'),
+                                            op_gauge_basis)
+            # Note: above, do we need to store identity? could we just use the basis as a space here? or 'None'?
+
+        # "reshape" mx so rows correspond to allowed_row_basis (the op's allowed labels)
+        # (maybe make this into a subroutine?)
+        assert(_sps.isspmatrix_csr(mx))
+        data = []; col_indices = []; rowptr = [0]  # build up a CSR matrix manually
+        allowed_lbls_set = set(allowed_row_basis.labels)
+        allowed_row_indices = [(i, allowed_row_basis.label_index(lbl))
+                               for i, lbl in enumerate(row_basis.labels) if lbl in allowed_lbls_set]
+
+        for i, new_i in sorted(allowed_row_indices, key=lambda x: x[1]):
+            # transfer i-th row of mx (whose rose are in row_basis) to new_i-th row of new mx
+
+            # - first increment rowptr as needed
+            while len(rowptr) <= new_i:
+                rowptr.append(len(data))
+
+            # - then add data
+            col_indices.extend(mx.indices[mx.indptr[i]:mx.indptr[i + 1]])
+            data.extend(mx.data[mx.indptr[i]:mx.indptr[i + 1]])
+            rowptr.append(len(data))
+        while len(rowptr) <= len(allowed_row_basis):  # fill in rowptr for any (empty) ending rows
+            rowptr.append(len(data))
+        allowed_rowspace_mx = _sps.csr_matrix((data, col_indices, rowptr),
+                                              shape=(len(allowed_row_basis), mx.shape[1]), dtype=mx.dtype)
+
+        return allowed_rowspace_mx, allowed_row_basis, op_gauge_space
+
+    def fogi_errorgen_component_labels(self, include_fogv=False, typ='normal'):
+        labels = self.fogi_store.fogi_errorgen_direction_labels(typ)
+        if include_fogv:
+            labels += self.fogi_store.fogv_errorgen_direction_labels(typ)
+        return labels
+
+    def fogi_errorgen_components_array(self, include_fogv=False, normalized_elem_gens=True):
+        op_coeffs = self.errorgen_coefficients(normalized_elem_gens)
+
+        if include_fogv:
+            fogi_coeffs, fogv_coeffs = self.fogi_store.opcoeffs_to_fogiv_components_array(op_coeffs)
+            return _np.concatenate((fogi_coeffs, fogv_coeffs))
+        else:
+            return self.fogi_store.opcoeffs_to_fogi_components_array(op_coeffs)
+
+    def set_fogi_errorgen_components_array(self, components, include_fogv=False, normalized_elem_gens=True,
+                                           truncate=False):
+        fogi, fogv = self.fogi_store.num_fogi_directions, self.fogi_store.num_fogv_directions
+
+        if include_fogv:
+            n = fogi
+            fogi_coeffs, fogv_coeffs = components[0:fogi], components[n: n + fogv]
+            op_coeffs = self.fogi_store.fogiv_components_array_to_opcoeffs(fogi_coeffs, fogv_coeffs)
+        else:
+            fogi_coeffs = components[0:fogi]
+            op_coeffs = self.fogi_store.fogi_components_array_to_opcoeffs(fogi_coeffs)
+
+        if not normalized_elem_gens:
+            def inv_rescale(coeffs):  # the inverse of the rescaling applied in fogi_errorgen_components_array
+                d2 = _np.sqrt(self.dim); d = _np.sqrt(d2)
+                return {lbl: (val * d if lbl.errorgen_type == 'H' else val) for lbl, val in coeffs.items()}
+        else:
+            def inv_rescale(coeffs): return coeffs
+
+        for op_label, coeff_dict in op_coeffs.items():
+            #TODO: update this conditional to something more robust (same conditiona in fogitools.py too)
+            if isinstance(op_label, str) and op_label.startswith('rho'):
+                self.preps[op_label].set_errorgen_coefficients(inv_rescale(coeff_dict), truncate=truncate)
+            elif isinstance(op_label, str) and op_label.startswith('M'):
+                self.povms[op_label].set_errorgen_coefficients(inv_rescale(coeff_dict), truncate=truncate)
+            else:
+                self.operations[op_label].set_errorgen_coefficients(inv_rescale(coeff_dict), truncate=truncate)
+
+    def fogi_errorgen_vector(self, normalized_elem_gens=False):
+        """
+        Constructs a vector from all the error generator coefficients involved in the FOGI analysis of this model.
+
+        Parameters
+        ----------
+        normalized_elem_gens : bool, optional
+            Whether or not coefficients correspond to elementary error generators
+            constructed from *normalized* Pauli matrices or not.
+
+        Returns
+        -------
+        numpy.ndarray
+        """
+        d = self.errorgen_coefficients(normalized_elem_gens=normalized_elem_gens)
+        errvec = _np.zeros(self.fogi_store.fogi_directions.shape[0], 'd')
+        for op_lbl in self.fogi_store.primitive_op_labels:
+            errdict = d[op_lbl]
+            elem_errgen_lbls = self.fogi_store.elem_errorgen_labels_by_op[op_lbl]
+            elem_errgen_indices = _slct.indices(self.fogi_store.op_errorgen_indices[op_lbl])
+            for elemgen_lbl, i in zip(elem_errgen_lbls, elem_errgen_indices):
+                errvec[i] = errdict.get(elemgen_lbl, 0.0)
+        return errvec
+
+    def _fogi_errorgen_vector_projection(self, space, normalized_elem_gens=False):
+        """ A helper function that projects self.errorgen_vector onto the space spanned by the columns of `space` """
+        errvec = self.fogi_errorgen_vector(normalized_elem_gens)
+        Pspace = space @ _np.linalg.pinv(space)  # construct projector
+        return Pspace @ errvec  # projected errvec
+
+    # create map parameter indices <=> fogi_vector_indices (for each fogi store)
+    def _create_model_parameter_to_fogi_errorgen_space_map(self):
+        fogi_store = self.fogi_store
+        num_elem_errgens, num_fogi_vecs = fogi_store.fogi_directions.shape
+        param_to_fogi_errgen_space_mx = _np.zeros((num_elem_errgens, self.num_params), 'd')
+        for op_label in fogi_store.primitive_op_labels:
+            elem_errgen_lbls = fogi_store.elem_errorgen_labels_by_op[op_label]
+            fogi_errgen_indices = _slct.indices(fogi_store.op_errorgen_indices[op_label])
+            assert(len(fogi_errgen_indices) == len(elem_errgen_lbls))
+
+            op = self.operations[op_label]
+            coeff_index_lookup = {elem_lbl: i for i, elem_lbl in enumerate(op.errorgen_coefficient_labels())}
+            coeff_indices = [coeff_index_lookup.get(elem_lbl, None) for elem_lbl in elem_errgen_lbls]
+
+            # For our particularly simple parameterization (H+s) op parameter indices == coeff indices:
+            assert(_np.allclose(op.errorgen_coefficients_array_deriv_wrt_params(), _np.identity(op.num_params))), \
+                "Currently only supported for simple parameterizations where op parameter indices == coeff indices"
+            op_param_indices = coeff_indices
+
+            gpindices = _slct.indices(op.gpindices)
+            mdl_param_indices = [(gpindices[i] if (i is not None) else None)
+                                 for i in op_param_indices]
+            for i_errgen, i_param in zip(fogi_errgen_indices, mdl_param_indices):
+                if i_param is not None:
+                    param_to_fogi_errgen_space_mx[i_errgen, i_param] = 1.0
+        return param_to_fogi_errgen_space_mx
+
+    def fogi_contribution(self, op_label, error_type='H', intrinsic_or_relational='intrinsic',
+                          target='all', hessian_for_errorbars=None):
+        """
+        Computes a contribution to the FOGI error on a single gate.
+
+        This method is used when partitioning the (FOGI) error on a gate in
+        various ways, based on the error type, whether the error is intrinsic
+        or relational, and the upon the error support.
+
+        Parameters
+        ----------
+        op_label : Label
+            The operation to compute a contribution for.
+
+        error_type : {'H', 'S', 'fogi_total_error', 'fogi_infidelity'}
+            The type of errors to include in the partition.  `'H'` means Hamiltonian
+            and `'S'` means Pauli stochastic.  There are two options for including
+            *both* H and S errors: `'fogi_total_error'` adds the Hamiltonian errors
+            linearly with the Pauli tochastic errors, similar to the diamond distance;
+            `'fogi_infidelity'` adds the Hamiltonian errors in quadrature to the linear
+            sum of Pauli stochastic errors, similar to the entanglement or average gate
+            infidelity.
+
+        intrinsic_or_relational : {"intrinsic", "relational", "all"}
+            Restrict to intrinsic or relational errors (or not, using `"all"`).
+
+        target : tuple or "all"
+            A tuple of state space (qubit) labels to restrict to, e.g., `('Q0','Q1')`.
+            Note that including multiple labels selects only those quantities that
+            target *all* the labels. The special `"all"` value includes quantities
+            on all targets (no restriction).
+
+        hessian_for_errorbars : numpy.ndarray, optional
+            If not `None`, a hessian matrix for this model (with shape `(Np, Np)`
+            where `Np == self.num_params`, the number of model paramters) that is
+            used to compute and return 1-sigma error bars.
+
+        Returns
+        -------
+        value : float
+            The value of the requested contribution.
+        errorbar : float
+            The 1-sigma error bar, returned *only* if `hessian_for_errorbars` is given.
+        """
+        if error_type in ('H', 'S'):
+            space = self.fogi_store.create_fogi_aggregate_single_op_space(op_label, error_type,
+                                                                          intrinsic_or_relational, target)
+            return self._fogi_contribution_single_type(error_type, space, hessian_for_errorbars)
+
+        elif error_type in ('fogi_total_error', 'fogi_infidelity'):
+            Hspace = self.fogi_store.create_fogi_aggregate_single_op_space(
+                op_label, 'H', intrinsic_or_relational, target)
+            Sspace = self.fogi_store.create_fogi_aggregate_single_op_space(
+                op_label, 'S', intrinsic_or_relational, target)
+            values = self._fogi_contribution_combined_HS_types(Hspace, Sspace, hessian_for_errorbars)
+            # (total, infidelity) if hessian is None otherwise (total, total_eb, infidelity, infidelity_eb)
+
+            if error_type == 'fogi_total_error':
+                return values[0] if (hessian_for_errorbars is None) else (values[0], values[1])
+            else:  # error_type == 'fogi_infidelity'
+                return values[1] if (hessian_for_errorbars is None) else (values[2], values[3])
+
+        else:
+            raise ValueError("Invalid error type: '%s'" % str(error_type))
+
+    def _fogi_contribution_single_type(self, errorgen_type, space, hessian=None):
+        """
+        Helper function to compute fogi contribution for a single error generator type,
+        where aggregation method is unambiguous.
+        Note: `space` should be a fogi-errgen-space subspace.
+        """
+        fogi_store = self.fogi_store
+        proj_errvec = self._fogi_errorgen_vector_projection(space, normalized_elem_gens=False)
+        if errorgen_type == 'H':
+            val = _np.linalg.norm(proj_errvec)
+        elif errorgen_type == 'S':
+            val = sum(proj_errvec)
+        else:
+            raise ValueError("Invalid `errorgen_type` '%s' - must be 'H' or 'S'!" % str(errorgen_type))
+
+        val = _np.real_if_close(val)
+        if abs(val) < 1e-10: val = 0.0
+
+        if hessian is not None:
+            if space.size == 0:  # special case
+                errbar = 0.0
+            else:
+                T = self._create_model_parameter_to_fogi_errorgen_space_map()
+                H_errgen_space = T @ hessian @ T.T
+
+                errgen_space_to_fogi = fogi_store.fogi_directions.toarray().T
+                pinv_espace_to_fogi = _np.linalg.pinv(errgen_space_to_fogi)
+                H_fogi = errgen_space_to_fogi @ H_errgen_space @ pinv_espace_to_fogi
+
+                inv_H_fogi = _np.linalg.pinv(H_fogi)  # hessian in fogi space
+
+                #convert fogi space back to errgen_space
+                inv_H_errgen_space = pinv_espace_to_fogi @ inv_H_fogi @ errgen_space_to_fogi
+
+                Pspace = space @ _np.linalg.pinv(space)
+                proj_inv_H_errgen_space = Pspace @ inv_H_errgen_space @ Pspace.T  # np.linalg.pinv(Pspace) #.T
+
+                if errorgen_type == 'H':
+                    # elements added in quadrature, val = sqrt( sum(element^2) ) = dot(proj_errvec_hat, proj_errvec)
+                    proj_errvec_hat = proj_errvec / _np.linalg.norm(proj_errvec)
+                    errbar = proj_errvec_hat.T @ proj_inv_H_errgen_space @ proj_errvec_hat
+                elif errorgen_type == "S":
+                    # elements added, val = sum(element) = dot(ones, proj_errvec)
+                    ones = _np.ones((proj_inv_H_errgen_space.shape[0], 1), 'd')
+                    errbar = ones.T @ proj_inv_H_errgen_space @ ones
+                else:
+                    raise ValueError("Invalid `errorgen_type`!")
+
+                if abs(errbar) < 1e-10: errbar = 0.0
+                errbar = _np.sqrt(float(_np.real_if_close(errbar)))
+
+        return val if (hessian is None) else (val, errbar)
+
+    def _fogi_contribution_combined_HS_types(self, Hspace, Sspace, hessian=None):
+        """
+        Helper function to compute fogi contribution for that combined multiple
+        (so far only works for H+S) error generator types, where there are multiple
+        aggregation methods (and all are computed).
+        Note: `space` should be a fogi-errgen-space subspace.
+        """
+        #TODO: maybe can combine with function above?
+        errvec = self.fogi_errorgen_vector(normalized_elem_gens=False)
+
+        Hvec = self._fogi_errorgen_vector_projection(Hspace, normalized_elem_gens=False)
+        Hhat = Hvec / _np.linalg.norm(Hvec)
+        Svec = _np.sum(Sspace, axis=1)  # should be all 1s and zeros
+        assert(all([(_np.isclose(el, 0) or _np.isclose(el, 1.0)) for el in Svec]))
+
+        total_error_vec = Hhat + Svec  # ~ concatenate((Hhat, Svec))
+        total_error_val = _np.dot(total_error_vec, errvec)
+
+        infidelity_vec = Hvec + Svec  # ~ concatenate((Hvec, Svec))
+        infidelity_val = _np.dot(infidelity_vec, errvec)
+
+        if hessian is not None:
+            T = self._create_model_parameter_to_fogi_errorgen_space_map()
+
+            hessian_errgen_space = T @ hessian @ T.T
+
+            errgen_space_to_fogi = self.fogi_store.fogi_directions.toarray().T
+            pinv_espace_to_fogi = _np.linalg.pinv(errgen_space_to_fogi)
+            H_fogi = errgen_space_to_fogi @ hessian_errgen_space @ pinv_espace_to_fogi
+
+            inv_H_fogi = _np.linalg.pinv(H_fogi)  # hessian in fogi space
+
+            #convert fogi space back to errgen_space
+            inv_H_errgen_space = pinv_espace_to_fogi @ inv_H_fogi @ errgen_space_to_fogi
+
+            total_error_eb = total_error_vec[None, :] @ inv_H_errgen_space @ total_error_vec[:, None]
+
+            infidelity_eb_vec = 2 * Hvec + Svec  # ~ concatenate((2 * Hvec, Svec))
+            infidelity_eb = infidelity_eb_vec[None, :] @ inv_H_errgen_space @ infidelity_eb_vec[:, None]
+
+            if abs(total_error_eb) < 1e-10: total_error_eb = 0.0
+            total_error_eb = _np.sqrt(float(_np.real_if_close(total_error_eb)))
+
+            if abs(infidelity_eb) < 1e-10: infidelity_eb = 0.0
+            infidelity_eb = _np.sqrt(float(_np.real_if_close(infidelity_eb)))
+
+            return total_error_val, total_error_eb, infidelity_val, infidelity_eb
+        else:
+            return total_error_val, infidelity_val
 
 
 def _default_param_bounds(num_params):
